Update usage.adoc
On my Windows 11 installation, using "Terminal (Admin)" I required quotes around the path. When I used exactly the commands used on this page, I got the following error:

```
 set PATH=C:\ProgramData\chocolatey\bin;C:\Windows\system32;C:\Windows
C:\Windows\system32 : The term 'C:\Windows\system32' is not recognized as the name of a cmdlet, function, script file,
or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and
try again.
At line:1 char:40
+ set PATH=C:\ProgramData\chocolatey\bin;C:\Windows\system32;C:\Windows
+                                        ~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : ObjectNotFound: (C:\Windows\system32:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException

C:\Windows : The term 'C:\Windows' is not recognized as the name of a cmdlet, function, script file, or operable
program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
At line:1 char:60
+ set PATH=C:\ProgramData\chocolatey\bin;C:\Windows\system32;C:\Windows
+                                                            ~~~~~~~~~~
    + CategoryInfo          : ObjectNotFound: (C:\Windows:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException

PS C:\Users\quint\Source\OpenCPN\weather_routing_pi> set PATH="C:\ProgramData\chocolatey\bin;C:\Windows\system32;C:\Windows"
```